requirements.txt
certifi==2018.11.29
chardet==3.0.4
falcon==1.4.1
gunicorn==19.9.0
idna==2.8
influxdb==5.2.1
numpy==1.16.0
pandas==0.23.4
patsy==0.5.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-mimeparse==1.6.0
pytz==2018.9
requests==2.21.0
scipy==1.2.0
six==1.12.0
statsmodels==0.9.0
urllib3==1.24.1
